fix(currency): Address review comments
- Move compact constructors into the formatter module (same type now)
- Store only the resolved usage instead of the full options bag
- Use the explicit standard negative pattern (essentials.get_negative)
  when formatting negative values in standard usage
- Preserve minus/plus signs when converting CLDR currency patterns in
  datagen (they were previously dropped) and regenerate baked data
