--- components/experimental/src/dimension/currency/compact_formatter.rs
@@ -1,499 +1,0 @@
-// This file is part of ICU4X. For terms of use, please see the file
-// called LICENSE at the top level of the ICU4X source tree
-// (online at: https://github.com/unicode-org/icu4x/blob/main/LICENSE ).
-
-use crate::dimension::provider::currency::{essentials::*, extended::*, patterns::*, symbols::*};
-use icu_decimal::CompactDecimalFormatter;
-use icu_provider::prelude::*;
-
-use super::CurrencyCode;
-use super::formatter::{CurrencyFormatter, CurrencyFormatterPreferences};
-use super::options::CurrencyFormatterOptions;
-
-impl CurrencyFormatter<CompactDecimalFormatter> {
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_symbol: skip,
-            try_new_compact_symbol_with_buffer_provider,
-            try_new_compact_symbol_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_symbol_narrow: skip,
-            try_new_compact_symbol_narrow_with_buffer_provider,
-            try_new_compact_symbol_narrow_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_code: skip,
-            try_new_compact_code_with_buffer_provider,
-            try_new_compact_code_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (prefs: CurrencyFormatterPreferences, currency_code: CurrencyCode) -> error: DataError,
-        functions: [
-            try_new_compact_name: skip,
-            try_new_compact_name_with_buffer_provider,
-            try_new_compact_name_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_long_symbol: skip,
-            try_new_compact_long_symbol_with_buffer_provider,
-            try_new_compact_long_symbol_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_long_symbol_narrow: skip,
-            try_new_compact_long_symbol_narrow_with_buffer_provider,
-            try_new_compact_long_symbol_narrow_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (
-            prefs: CurrencyFormatterPreferences,
-            currency_code: CurrencyCode,
-            options: CurrencyFormatterOptions
-        ) -> error: DataError,
-        functions: [
-            try_new_compact_long_code: skip,
-            try_new_compact_long_code_with_buffer_provider,
-            try_new_compact_long_code_unstable,
-            Self
-        ]
-    );
-
-    icu_provider::gen_buffer_data_constructors!(
-        (prefs: CurrencyFormatterPreferences, currency_code: CurrencyCode) -> error: DataError,
-        functions: [
-            try_new_compact_long_name: skip,
-            try_new_compact_long_name_with_buffer_provider,
-            try_new_compact_long_name_unstable,
-            Self
-        ]
-    );
-
-    /// Creates a new [`CurrencyFormatter`] for compact short number formatting with short currency symbols from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_symbol(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_essential(
-            CompactDecimalFormatter::try_new_short((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::SHORT,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact short number formatting with narrow currency symbols from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_symbol_narrow(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_essential(
-            CompactDecimalFormatter::try_new_short((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::NARROW,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact short number formatting using the 3-letter ISO currency code from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_code(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_code_internal(
-            CompactDecimalFormatter::try_new_short((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact short number formatting with full currency display names from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_name(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-    ) -> Result<Self, DataError> {
-        Self::try_new_name_internal(
-            CompactDecimalFormatter::try_new_short((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact long number formatting with short currency symbols from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_long_symbol(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_essential(
-            CompactDecimalFormatter::try_new_long((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::SHORT,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact long number formatting with narrow currency symbols from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_long_symbol_narrow(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_essential(
-            CompactDecimalFormatter::try_new_long((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::NARROW,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact long number formatting using the 3-letter ISO currency code from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_long_code(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError> {
-        Self::try_new_code_internal(
-            CompactDecimalFormatter::try_new_long((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-            options,
-        )
-    }
-
-    /// Creates a new [`CurrencyFormatter`] for compact long number formatting with full currency display names from compiled locale data.
-    ///
-    /// ✨ *Enabled with the `compiled_data` Cargo feature.*
-    ///
-    /// [📚 Help choosing a constructor](icu_provider::constructors)
-    #[cfg(feature = "compiled_data")]
-    pub fn try_new_compact_long_name(
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-    ) -> Result<Self, DataError> {
-        Self::try_new_name_internal(
-            CompactDecimalFormatter::try_new_long((&prefs).into(), Default::default())?,
-            prefs,
-            currency_code,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_symbol)]
-    pub fn try_new_compact_symbol_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<CurrencySymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactShortV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_essential_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_short_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::SHORT,
-            options,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_symbol_narrow)]
-    pub fn try_new_compact_symbol_narrow_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<CurrencySymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactShortV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_essential_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_short_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::NARROW,
-            options,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_name)]
-    pub fn try_new_compact_name_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyExtendedDataV1>
-            + DataProvider<CurrencyPatternsDataV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactShortV1>,
-    {
-        Self::try_new_name_internal_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_short_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_long_symbol)]
-    pub fn try_new_compact_long_symbol_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<CurrencySymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactLongV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_essential_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_long_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::SHORT,
-            options,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_long_symbol_narrow)]
-    pub fn try_new_compact_long_symbol_narrow_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<CurrencySymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactLongV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_essential_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_long_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            CurrencySymbolsV1::NARROW,
-            options,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_long_name)]
-    pub fn try_new_compact_long_name_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyExtendedDataV1>
-            + DataProvider<CurrencyPatternsDataV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactLongV1>,
-    {
-        Self::try_new_name_internal_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_long_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_code)]
-    pub fn try_new_compact_code_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactShortV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_code_internal_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_short_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            options,
-        )
-    }
-
-    #[doc = icu_provider::gen_buffer_unstable_docs!(UNSTABLE, Self::try_new_compact_long_code)]
-    pub fn try_new_compact_long_code_unstable<D>(
-        provider: &D,
-        prefs: CurrencyFormatterPreferences,
-        currency_code: CurrencyCode,
-        options: CurrencyFormatterOptions,
-    ) -> Result<Self, DataError>
-    where
-        D: ?Sized
-            + DataProvider<CurrencyEssentialsV1>
-            + DataProvider<icu_decimal::provider::DecimalCompactLongV1>
-            + DataProvider<icu_decimal::provider::DecimalSymbolsV1>
-            + DataProvider<icu_decimal::provider::DecimalDigitsV1>
-            + DataProvider<icu_plurals::provider::PluralsCardinalV1>,
-    {
-        Self::try_new_code_internal_unstable(
-            provider,
-            CompactDecimalFormatter::try_new_long_unstable(
-                provider,
-                (&prefs).into(),
-                Default::default(),
-            )?,
-            prefs,
-            currency_code,
-            options,
-        )
-    }
-}
